refactor: create RouterContext
